refactor: fix eslint issues
