--- src/app.js
@@ -1,168 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import { connect } from 'react-redux';
-
-import {
-  getAggregateYearData,
-  getAggregateStateData,
-  getAggregateLocationData,
-  getAvailableStates,
-  getPredictions,
-  getUserFromStorage,
-  setChartMode,
-  setDataMode,
-} from './state/actions';
-
-import {
-  About,
-  Admin,
-  TrappingData,
-  Home,
-  PlayWithModel,
-  Prediction,
-} from './screens';
-
-import {
-  Header,
-  Footer,
-  MobileOverlay,
-  ScrollToTop,
-} from './components';
-
-import {
-  CHART_MODES,
-  DATA_MODES,
-  getAutomationServerUrl,
-  getServerUrl,
-  MIN_WIDTH_THRESHOLD,
-  ROUTES,
-} from './constants';
-
-import {
-  getAuthTokenFromStorage,
-  getDataModeFromStorage,
-  getUserIdFromStorage,
-} from './utils';
-
-const FallBack = () => {
-  return <div>URL not found</div>;
-};
-
-global.API_URL = getServerUrl();
-global.AUTOMATION_API_URL = getAutomationServerUrl();
-
-const App = (props) => {
-  const {
-    loginUserFromStorage,
-    predictionYear,
-    startYear,
-    chartMode,
-  } = props;
-
-  const [isMobile, setIsMobile] = useState(window.innerWidth < MIN_WIDTH_THRESHOLD);
-
-  useEffect(() => {
-    global.API_URL = getServerUrl();
-    global.AUTOMATION_API_URL = getAutomationServerUrl();
-
-    // fetch user data if persist in browser
-    if (getAuthTokenFromStorage() && getUserIdFromStorage()) {
-      loginUserFromStorage();
-    }
-
-    // set data/chart mode if persist in browser
-    props.setDataMode(getDataModeFromStorage() || DATA_MODES.COUNTY);
-    props.setChartMode(CHART_MODES.MAP);
-
-    // fetch initial data
-    props.getAggregateYearData();
-    props.getAggregateStateData();
-    props.getAggregateLocationData();
-
-    const resizeListener = (e) => setIsMobile(e.target.innerWidth < MIN_WIDTH_THRESHOLD);
-    window.addEventListener('resize', resizeListener);
-
-    return () => window.removeEventListener('resize', resizeListener);
-  }, []);
-
-  useEffect(() => {
-    if (chartMode === CHART_MODES.GRAPH) {
-      props.getPredictions(startYear, predictionYear);
-      props.getAvailableStates({ predictionYear });
-    } else {
-      props.getPredictions(predictionYear, predictionYear);
-      props.getAvailableStates({ predictionYear });
-    }
-  }, [startYear, predictionYear, chartMode]);
-
-  if (isMobile) return <MobileOverlay />;
-
-  return (
-    <Router>
-      <ScrollToTop>
-        <div>
-          <Header />
-          <div className="content">
-            <Switch>
-              <Route exact path={ROUTES.HOME} component={Home} />
-              <Route path={ROUTES.ABOUT} component={About} />
-              <Route path={ROUTES.ADMIN} component={Admin} />
-              <Route path={ROUTES.TRAPPING_DATA} component={TrappingData} />
-              <Route path={ROUTES.PLAY_WITH_MODEL} component={PlayWithModel} />
-              <Route path={ROUTES.PREDICTIONS} component={Prediction} />
-              <Route component={FallBack} />
-            </Switch>
-          </div>
-          <Footer />
-        </div>
-      </ScrollToTop>
-    </Router>
-  );
-};
-
-const mapStateToProps = (state) => {
-  const {
-    selections: {
-      startYear,
-      predictionYear,
-      chartMode,
-    },
-  } = state;
-
-  return {
-    startYear,
-    predictionYear,
-    chartMode,
-  };
-};
-
-const mapDispatchToProps = (dispatch) => {
-  return {
-    getAggregateYearData: () => {
-      dispatch(getAggregateYearData());
-    },
-    getAggregateStateData: () => {
-      dispatch(getAggregateStateData());
-    },
-    getAggregateLocationData: () => {
-      dispatch(getAggregateLocationData());
-    },
-    getAvailableStates: (overrideFilter) => {
-      dispatch(getAvailableStates(overrideFilter));
-    },
-    getPredictions: (startYear, endYear) => {
-      dispatch(getPredictions(startYear, endYear));
-    },
-    loginUserFromStorage: () => {
-      dispatch(getUserFromStorage());
-    },
-    setChartMode: (mode) => {
-      dispatch(setChartMode(mode));
-    },
-    setDataMode: (mode) => {
-      dispatch(setDataMode(mode));
-    },
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
